# .plans/core/api/openapi.yaml
openapi: 3.0.0
info:
  title: HG Conversion API
  version: 1.0.0
paths:
  /convert/units:
    post:
      summary: Converte um valor entre duas unidades
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                value:
                  type: number
                  description: Valor numérico a converter
                from:
                  type: string
                  description: Unidade de origem
                to:
                  type: string
                  description: Unidade de destino
                analyte:
                  type: string
                  description: Identificador do analito (opcional)
              required: [value, from, to]
      responses:
        '200':
          description: Resultado da conversão
          content:
            application/json:
              schema:
                type: object
                properties:
                  value:
                    type: number
                  unit:
                    type: string
                  norm:
                    type: object
                    description: Informações normalizadas do analito e unidades
  /calculators:
    get:
      summary: Lista as calculadoras disponíveis
      responses:
        '200':
          description: Lista de calculadoras
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    version:
                      type: string
  /calculators/{id}:
    get:
      summary: Obtém o esquema de uma calculadora específica
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Esquema da calculadora
          content:
            application/json:
              schema:
                type: object
    post:
      summary: Executa cálculos de uma calculadora
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                inputs:
                  type: object
                  description: Valores de entrada conforme especificação da calculadora
              required: [inputs]
      responses:
        '200':
          description: Resultado do cálculo
          content:
            application/json:
              schema:
                type: object
                properties:
                  outputs:
                    type: object
                  explanation:
                    type: string
                    description: Representação textual das fórmulas usadas